refactor: move Metal-specific codegen into `helion/_compiler/metal/`
Relocate the Metal backend `@_decorators.codegen(op, "metal")` bodies out of the
backend-neutral language modules into the Metal backend folder:

  `helion/_compiler/metal/metal_tracing_ops.py`  (`_mask_to`)
  `helion/_compiler/metal/metal_memory_ops.py`   (`store`, `load`)

Each language module keeps the backend-neutral op + a bottom
`from .._compiler.metal import metal_<module>  # noqa: E402, F401` (ruff sorts the
backend imports cute/metal/pallas/triton). No metal-only module-level helpers
existed, so none were moved; shared imports stay in the language module.

No behavior change: all 3 metal codegens now resolve from
`helion._compiler.metal.*`; no residual metal codegens in `helion/language/`;
other backends untouched; ruff clean. This completes the per-backend codegen
separation for all backends (pallas, triton, cute, metal).

stack-info: PR: https://github.com/pytorch/helion/pull/3052, branch: AmesingFlank/stack/71
