+++ helion/language/memory_ops.py
@@ -4443,30 +4443,6 @@ def _codegen_cute_store_permute_lane_loops(
     )
 
 
-@_decorators.codegen(store, "metal")
-def _(state: CodegenState) -> ast.AST:
-    # Metal delegates to the same PointerIndexingStrategy as Triton.
-    # This produces tl.store(ptr + offset, val, mask) in the AST;
-    # the MSL walker translates it to Metal.
-    tensor = state.proxy_arg(0)
-    subscript = state.proxy_arg(1)
-    assert isinstance(subscript, (list, tuple))
-    value = state.ast_arg(2)
-    extra_mask = state.ast_args[3]
-    assert isinstance(extra_mask, (type(None), ast.AST))
-
-    if isinstance(tensor, torch.Tensor):
-        device_fn = state.device_function
-        device_fn.device_store_index += 1
-        indexing_idx = device_fn.device_memory_op_index
-        device_fn.device_memory_op_index += 1
-        strategy = device_fn.get_indexing_strategy(indexing_idx)
-        return strategy.codegen_store(
-            state, tensor, [*subscript], value, extra_mask, None
-        )
-    raise exc.BackendUnsupported("metal", f"store target type: {type(tensor)}")
-
-
 # TODO(joydddd): Add support for stack tensor in ref mode.
 @_decorators.ref(store)
 def _(
@@ -4656,33 +4632,6 @@ def _maybe_materialize_tile_index_load(
     return expr_from_string(f"{base_var}[{', '.join(parts)}]")
 
 
-@_decorators.codegen(load, "metal")
-def _(state: CodegenState) -> ast.AST:
-    # Metal delegates to the same PointerIndexingStrategy as Triton.
-    # This produces tl.load(ptr + offset, mask, other=0) in the AST;
-    # the MSL walker translates it to Metal.
-    tensor = state.proxy_arg(0)
-    subscript = state.proxy_arg(1)
-    assert isinstance(subscript, (list, tuple))
-    ast_subscript = state.ast_args[1]
-    assert isinstance(ast_subscript, (list, tuple))
-    extra_mask = state.ast_args[2]
-    assert isinstance(extra_mask, (type(None), ast.AST))
-    eviction_policy = state.ast_args[3] if len(state.ast_args) > 3 else None
-    assert isinstance(eviction_policy, (type(None), ast.AST))
-
-    if isinstance(tensor, torch.Tensor):
-        device_fn = state.device_function
-        device_fn.device_load_index += 1
-        indexing_idx = device_fn.device_memory_op_index
-        device_fn.device_memory_op_index += 1
-        strategy = device_fn.get_indexing_strategy(indexing_idx)
-        return strategy.codegen_load(
-            state, tensor, [*subscript], extra_mask, eviction_policy, None
-        )
-    raise exc.BackendUnsupported("metal", f"load tensor type: {type(tensor)}")
-
-
 @_decorators.get_masked_value(load)
 def _(node: torch.fx.Node) -> int:
     return 0  # loads are always masked to 0
@@ -4785,5 +4734,6 @@ def _(
 # @_decorators.codegen(op, "<backend>") registrations run with the same eager
 # timing as when the bodies lived in this file -- no behavior change.
 import helion._compiler.cute.memory_ops  # noqa: E402, F401
+import helion._compiler.metal.memory_ops  # noqa: E402, F401
 import helion._compiler.pallas.memory_ops  # noqa: E402, F401
 import helion._compiler.triton.memory_ops  # noqa: E402, F401
